Find the User's Socket
Small change to TellUserAboutContactstatusChange.
If there is no socket in the args, check the _uesrs list for one.

This event is only fired when a user goes offline right now anyway, so
it doesn't matter much.

I should make the server simply manage online status and sockets only in
the memory, and not in the database.

Minor:
- Renamed "UserStatusChangeEventArgs" to the same standard as the rest
of the code.