Failed Login Attempts
Created a dictionary, "_tempUser" , to store the sharedSecret of a user
if the login attempt failed.
Not super happy with how it works, but at least it means that client
doesn't need changing.

It could cause problems if multiple login attempts are made from the
same client at the same time. But hopefully that wouldn't ever happen
unless someone modifies their client.